PDF-TTS-CONVERTER/requirements.txt
Flask==2.1.1
PyPDF2==3.0.1
gTTS==2.3.2
Werkzeug==2.1.1
gunicorn
pymongo~=4.7.3
python-dotenv
pytesseract
pdf2image
Pillow
